refactor: change project structure with preflight, utils and globals packages
