+++ nuv/utils/util_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 //
-package main
+package utils
 
 import "fmt"
